chore(examples): rename non-route files with underscore (#212)
