+++ packages/react-server/examples/basic/src/routes/demo/waku_02/page.tsx
@@ -1,6 +1,6 @@
 import { Link } from "@hiogawa/react-server/client";
 import { sortBy } from "@hiogawa/utils";
-import { fetchPokemons } from "./utils";
+import { fetchPokemons } from "./_utils";
 
 export default async function Page() {
   const pokemons = await fetchPokemons();
