+++ packages/react-server/examples/basic/src/routes/test/other/_client.tsx
