+++ packages/react-server/examples/starter/src/routes/server-action/page.tsx
@@ -1,5 +1,5 @@
-import { Counter } from "./client";
-import { getCounter } from "./server";
+import { getCounter } from "./_action";
+import { Counter } from "./_client";
 
 export default async function Page() {
   return (
