+++ packages/react-server/examples/starter/src/routes/use-state/page.tsx
@@ -1,4 +1,4 @@
-import { Counter } from "./client";
+import { Counter } from "./_client";
 
 export default function Page() {
   return (
